Seperating out the push client from window initialisation to try and make testing easier
